@@ -1,3 +1,4 @@
+import open from 'open';
 import { Config } from '../config';
 import { findXcodePath } from './common';
 import { wait } from '../common';
@@ -6,8 +7,7 @@ export async function openIOS(config: Config) {
   const xcodeProject = await findXcodePath(config);
 
   if (xcodeProject) {
-    const opn = await import('open');
-    await opn(xcodeProject, { wait: false });
+    await open(xcodeProject, { wait: false });
     await wait(3000);
   } else {
     throw new Error(
