@@ -12,9 +12,9 @@ import { updateCommand } from '../src/tasks/update';
 describe.each([false, true])(
   'Update: Android (monoRepoLike: %p)',
   monoRepoLike => {
-    let appDirObj;
-    let appDir;
-    let FS;
+    let appDirObj: any;
+    let appDir: string;
+    let FS: MappedFS;
 
     beforeAll(async () => {
       // These commands are slowww...
